# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PayReality Runtime Authority API",
    "description": "Deterministic policy evaluation (OPA/Rego), ED25519-signed Evidence, and the human-review resolution flow for AI agent financial actions. Full schema at /openapi.json, interactive docs at /docs.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/policies/documents": {
      "get": {
        "tags": [
          "policies"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_v1_policies_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "title": "Response List Documents V1 Policies Documents Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "policies"
        ],
        "summary": "Upload Document",
        "description": "spec 19.1. Extraction (spec 12.4 Stage 2-3) runs synchronously here\nfor Phase 1 simplicity: a document transitions extraction_pending ->\nextracted|extraction_failed within this same request.",
        "operationId": "upload_document_v1_policies_documents_post",
        "parameters": [
          {
            "name": "x-payreality-operator-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Payreality-Operator-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_v1_policies_documents_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/authorities": {
      "get": {
        "tags": [
          "policies"
        ],
        "summary": "List Authorities",
        "description": "spec 19.2.",
        "operationId": "list_authorities_v1_policies_authorities_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Document Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthorityResponse"
                  },
                  "title": "Response List Authorities V1 Policies Authorities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/authorities/{authority_id}": {
      "patch": {
        "tags": [
          "policies"
        ],
        "summary": "Review Authority",
        "description": "spec 19.2 / Section 13.",
        "operationId": "review_authority_v1_policies_authorities__authority_id__patch",
        "parameters": [
          {
            "name": "authority_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Authority Id"
            }
          },
          {
            "name": "x-payreality-operator-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Payreality-Operator-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewAuthorityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/{document_id}/compile": {
      "post": {
        "tags": [
          "policies"
        ],
        "summary": "Compile Policy",
        "description": "spec 19.3.",
        "operationId": "compile_policy_v1_policies__document_id__compile_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "x-payreality-operator-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Payreality-Operator-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompilePolicyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/{policy_id}/activate": {
      "post": {
        "tags": [
          "policies"
        ],
        "summary": "Activate Policy",
        "description": "spec 19.3 + 14.4 (also used for rollback: reactivating a retired\nversion's id).",
        "operationId": "activate_policy_v1_policies__policy_id__activate_post",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Id"
            }
          },
          {
            "name": "x-payreality-operator-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Payreality-Operator-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivatePolicyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivatePolicyResponse": {
        "properties": {
          "policy_id": {
            "type": "string",
            "format": "uuid",
            "title": "Policy Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "activated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At"
          },
          "previous_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Version"
          }
        },
        "type": "object",
        "required": [
          "policy_id",
          "version",
          "status",
          "activated_at",
          "previous_version"
        ],
        "title": "ActivatePolicyResponse"
      },
...
      "AuthorityResponse": {
        "properties": {
          "authority_id": {
            "type": "string",
            "format": "uuid",
            "title": "Authority Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "principal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Principal Id"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "limit_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Amount"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "conditions": {
            "items": {},
            "type": "array",
            "title": "Conditions"
          },
          "source_excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Excerpt"
          },
          "source_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Page"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "validation_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Validation Flags",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "authority_id",
          "document_id",
          "principal_id",
          "scope",
          "limit_amount",
          "currency",
          "conditions",
          "source_excerpt",
          "source_page",
          "status",
          "reviewer_id",
          "rejection_reason"
        ],
        "title": "AuthorityResponse"
      },
      "Body_upload_document_v1_policies_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_v1_policies_documents_post"
      },
      "CompilePolicyResponse": {
        "properties": {
          "policy_id": {
            "type": "string",
            "format": "uuid",
            "title": "Policy Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "bundle_hash": {
            "type": "string",
            "title": "Bundle Hash"
          },
          "mandate_count": {
            "type": "integer",
            "title": "Mandate Count"
          }
        },
        "type": "object",
        "required": [
          "policy_id",
          "version",
          "status",
          "bundle_hash",
          "mandate_count"
        ],
        "title": "CompilePolicyResponse"
      },
...
      "DocumentResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "name",
          "status",
          "uploaded_at"
        ],
        "title": "DocumentResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewAuthorityRequest": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reviewer_id": {
            "type": "string",
            "title": "Reviewer Id"
          },
          "edits": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edits"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          }
        },
        "type": "object",
        "required": [
          "status",
          "reviewer_id"
        ],
        "title": "ReviewAuthorityRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}